Move spatial mapping scripts to Scripts/
