+++ unity/Hello_World/Assets/Scripts/SpatialMappingCapture.cs.meta
